refactor: change modal component to form page
