--- src/components/modal/modal.Interface.ts
@@ -1,3 +1,0 @@
-export interface IModal {
-  isModalOpen: boolean;
-}
