+++ src/pages/form/form.Interface.ts
@@ -1,0 +1,6 @@
+export interface IClassTag {
+  selected?: boolean;
+  title?: string;
+  loading?: boolean;
+  onClick?: (e: React.MouseEvent<HTMLElement>) => void;
+}
