+++ src/pages/form/form.style.scss
@@ -7,6 +7,7 @@
   top: 0;
   left: 0;
   right: 0;
+  min-height: 100vh;
   z-index: 3;
   display: flex;
   justify-content: center;
@@ -85,6 +86,7 @@
     }
 
     &__tag {
+      min-width: 72px;
       font-size: 0.85rem;
       user-select: none;
       background-color: var.$blue-darker;
@@ -92,6 +94,7 @@
       padding: 0.5rem 0.75rem;
       display: flex;
       align-items: center;
+      justify-content: center;
       border-radius: 5px;
       box-shadow: 0 0 5px -1px black;
       letter-spacing: 1px;
